Refactor : cam 모듈과 cams 모듈 병합
서로 의존성이 매우 크고, 비슷한 네이밍으로 혼란을 줄 수 있기 때문에 두 모듈을 병합하였습니다.
