+++ backend/src/cam/cam.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CamController } from './cam.controller';
-import { CamService } from './cam.service';
+import { CamInnerService } from './cam-inner.service';
 
 describe('CamController', () => {
   let controller: CamController;
@@ -8,7 +8,7 @@ describe('CamController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [CamController],
-      providers: [CamService],
+      providers: [CamInnerService],
     }).compile();
 
     controller = module.get<CamController>(CamController);
