+++ backend/src/cam/cam.entity.ts
@@ -8,7 +8,7 @@ import {
 import { Server } from '../server/server.entity';
 
 @Entity()
-export class Cams {
+export class Cam {
   @PrimaryGeneratedColumn({ type: 'bigint' })
   id: number;
 
@@ -21,6 +21,6 @@ export class Cams {
   @ManyToOne(() => Server)
   server: Server;
 
-  @RelationId((cams: Cams) => cams.server)
+  @RelationId((cam: Cam) => cam.server)
   serverId: number;
 }
