+++ backend/src/cam/cam.gateway.spec.ts
@@ -1,18 +1,18 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CamGateway } from './cam.gateway';
-import { CamService } from './cam.service';
+import { CamInnerService } from './cam-inner.service';
 
 describe('CamGateway', () => {
   let gateway: CamGateway;
-  let service: CamService;
+  let service: CamInnerService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [CamGateway, CamService],
+      providers: [CamGateway, CamInnerService],
     }).compile();
 
     gateway = module.get<CamGateway>(CamGateway);
-    service = module.get<CamService>(CamService);
+    service = module.get<CamInnerService>(CamInnerService);
   });
 
   it('should be defined', () => {
