+++ backend/src/cam/cam.repository.ts
@@ -1,0 +1,13 @@
+import { EntityRepository, Repository } from 'typeorm';
+import { Cam } from './cam.entity';
+
+@EntityRepository(Cam)
+export class CamRepository extends Repository<Cam> {
+  findByServerId(serverId: number) {
+    return this.createQueryBuilder('cam')
+      .where('cam.serverId = :serverId', {
+        serverId,
+      })
+      .getMany();
+  }
+}
