+++ frontend/src/components/Main/Cam/CreateCamModal.tsx
@@ -180,7 +180,7 @@ function CreateCamModal(): JSX.Element {
 
   const onSubmitCreateCamModal = async (data: { name: string; description: string }) => {
     const { name } = data;
-    await fetch('api/cams', {
+    await fetch('api/cam', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
